refactor: component 폴더 정리 및 불필요한 component 제거
