+++ frontend/src/pages/LoginPage/components/Logo.tsx
