+++ frontend/src/pages/MainPage/components/MyTeamCard.tsx
