+++ frontend/src/pages/MyPage/components/ReceivedRollingpaperCard.tsx
