+++ frontend/src/pages/MyPage/components/WrittenMessageCard.tsx
