+++ frontend/src/pages/RollingpaperPage/components/MessageColorPicker.tsx
