+++ frontend/src/pages/TeamDetailPage/components/RollingpaperListItem.tsx
