+++ frontend/src/pages/TeamDetailPage/components/TeamRollingpaperList.tsx
@@ -5,7 +5,7 @@ import axios from "axios";
 import styled from "@emotion/styled";
 
 import IconButton from "@/components/IconButton";
-import RollingpaperListItem from "@/components/RollingpaperListItem";
+import RollingpaperListItem from "@/pages/TeamDetailPage/components/RollingpaperListItem";
 
 import appClient from "@/api";
 import { CustomError } from "@/types";
