+++ frontend/src/pages/TeamSearchPage/components/TeamCard.tsx
